+++ src/client/useCurrentLocale.ts
@@ -3,8 +3,43 @@
 import { Config } from '../types';
 import { usePathname } from 'next/navigation';
 
-const useCurrentLocale = (i18nConfig: Config): string | undefined => {
+const getCookie = (name: string, documentCookie: string) => {
+  let cookies = documentCookie;
+  let cookieArray = cookies.split('; ');
+
+  for (let i = 0; i < cookieArray.length; i++) {
+    let cookiePair = cookieArray[i].split('=');
+    if (cookiePair[0] === name) {
+      return cookiePair[1];
+    }
+  }
+  return null;
+};
+
+const useCurrentLocale = (
+  i18nConfig: Config,
+  documentCookie?: string
+): string | undefined => {
   const { basePath = '', locales } = i18nConfig;
+
+  if (!documentCookie && typeof window !== 'undefined') {
+    documentCookie = document.cookie;
+  }
+
+  if (documentCookie) {
+    const cookieLocale = getCookie(
+      i18nConfig.localeCookie || 'NEXT_LOCALE',
+      documentCookie
+    );
+
+    if (cookieLocale && locales.includes(cookieLocale)) {
+      return cookieLocale;
+    }
+  }
+
+  if (i18nConfig.noPrefix) {
+    return i18nConfig.defaultLocale;
+  }
 
   const currentPathname = usePathname();