+++ src/types.ts
@@ -6,6 +6,7 @@ export interface Config {
   localeCookie?: string;
   localeDetector?: ((request: NextRequest, config: Config) => string) | false;
   prefixDefault?: boolean;
+  noPrefix?: boolean;
   basePath?: string;
   serverSetCookie?: 'if-empty' | 'always' | 'never';
 }